Moved GitHub repo creation to separate construct
